Upgrade PHPStan to 2.x, clean codebase to level 8, and fix bugs

+++ src/Console/Commands/Cache/TtlCommand.php
@@ -175,7 +175,7 @@ class TtlCommand extends BaseCommand
         $units = ['B', 'KB', 'MB', 'GB'];
         $bytes = max($bytes, 0);
         $pow = floor(($bytes > 0 ? log($bytes) : 0) / log(1024));
-        $pow = min($pow, count($units) - 1);
+        $pow = (int) min($pow, count($units) - 1);
         $bytes /= (1024 ** $pow);
         return round($bytes, 2) . ' ' . $units[$pow];
     }
