+++ src/Performance/MemoryManager.php
@@ -188,7 +188,7 @@ class MemoryManager
         $units = ['B', 'KB', 'MB', 'GB', 'TB', 'PB'];
         $bytes = max($bytes, 0);
         $pow = floor(log($bytes) / log(1024));
-        $pow = min($pow, count($units) - 1);
+        $pow = (int) min($pow, count($units) - 1);
 
         $bytes /= pow(1024, $pow);
